restore y forceDelete + Gate y Policy

+++ app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use App\Events\ProjectSaved;
 use App\Http\Requests\SaveProjectRequest;
 use App\Models\Category;
 use App\Models\Project;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class ProjectController extends Controller
@@ -18,15 +19,20 @@ class ProjectController extends Controller
 
     public function index()
     {
+        // return Project::onlyTrashed()->with('category')->latest()->paginate();
         $projects = Project::with('category')->latest()->paginate();
+        $newProject = new Project;
+        $deletedProjects =  Project::onlyTrashed()->get();
 
-        return view('projects.index', compact('projects'));
+        return view('projects.index', compact('projects', 'newProject', 'deletedProjects'));
     }
 
     public function create()
     {
+        $this->authorize('create', $project = new Project);
+
         return view('projects.create', [
-            'project' => new Project(),
+            'project' => $project,
             'categories' => Category::pluck('name', 'id'),
         ]);
     }
@@ -34,7 +40,11 @@ class ProjectController extends Controller
     public function store(SaveProjectRequest $request)
     {
         $project = new Project( $request->validated() );
+        
+        $this->authorize('create', $project);
+
         $project->image = $request->file('image')->store('images', 'public');
+
         $project->save();
 
         ProjectSaved::dispatch($project);
@@ -49,6 +59,8 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
+        $this->authorize('update', $project);
+
         return view('projects.edit', [
             'project'    => $project,
             'categories' => Category::pluck('name', 'id'),
@@ -57,6 +69,8 @@ class ProjectController extends Controller
 
     public function update(SaveProjectRequest $request, Project $project)
     {
+        $this->authorize('update', $project);
+
         if ( $request->hasFile('image') ) {
             Storage::disk('public')->delete($project->image);
 
@@ -76,10 +90,36 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
-        Storage::disk('public')->delete($project->image);
+        $this->authorize('delete', $project);
+
+        // Storage::disk('public')->delete($project->image);
 
         $project->delete();
         
         return redirect()->route('projects.index')->with('status', 'El proyecto fue eliminado con éxito');
+    }
+
+    public function restore($projectUrl)
+    {
+        $project = Project::withTrashed()->whereUrl($projectUrl)->firstOrFail();
+
+        $this->authorize('restore', $project);
+
+        $project->restore();
+        
+        return redirect()->route('projects.index')->with('status', 'El proyecto fue eliminado restaurado con éxito');
+    }
+
+    public function forceDelete($projectUrl)
+    {
+        $project = Project::withTrashed()->whereUrl($projectUrl)->firstOrFail();
+
+        $this->authorize('forceDelete', $project);
+
+        Storage::disk('public')->delete($project->image);
+        
+        $project->forceDelete();
+        
+        return redirect()->route('projects.index')->with('status', 'El proyecto fue eliminado permanentemente');
     }
 }
